chore: extract proxy context parsing into its own home
